refactor: RacingGameTest 개선 및 OutputView 구조개선
OutputView 자동차의 이동경로 출력 및 경주 승자 출력 로직 구현
RacingGameTest TestMoveableStrategy를 바탕으로 이동,승자 테스트 구현
CarsTest Cars의 작동에 관한 테스트 구현
